+++ src/components/Item.js
@@ -8,13 +8,14 @@
  * @return {Element} - The created result item element
  */
 export default (data, index, config) => {
-  // Create a DIV element for each matching result item
   const item = document.createElement(config.resultItem.element);
+  // Setting item element attributes
   item.setAttribute("id", `${config.resultItem.idName}_${index}`);
   item.setAttribute("class", config.resultItem.className);
   item.setAttribute("role", "option");
+
   item.innerHTML = data.match;
-  // If custom content set pass params
+  // If custom content is active pass params
   if (config.resultItem.content) config.resultItem.content(data, item);
 
   return item;