+++ src/controllers/dataController.js
@@ -1,9 +1,9 @@
 import searchEngine from "../services/search";
 
 /**
- * Find matches to `query`
+ * Find matches to "query"
  *
- * @param {Object} config - The search engine configurations
+ * @param {Object} config - Search engine configurations
  * @param {String} query - User's search query string
  *
  * @return {Array} - Matches
@@ -13,15 +13,16 @@ export default (config, query) => {
 
   const results = [];
 
-  // Get matches in data source
+  // Find matches from data source
   data.store.forEach((record, index) => {
     const search = (key) => {
       const recordValue = (key ? record[key] : record).toString();
-      if (!recordValue) return;
-      
-      const match = typeof customSearchEngine === "function"
+
+      const match =
+        typeof customSearchEngine === "function"
           ? customSearchEngine(query, recordValue)
           : searchEngine(query, recordValue, config);
+
       if (!match) return;
 
       let result = {
@@ -29,9 +30,12 @@ export default (config, query) => {
         match,
         value: record,
       };
-      if (key) result.key = key
+
+      if (key) result.key = key;
+
       results.push(result);
     };
+
     if (data.key) {
       for (const key of data.key) {
         search(key);